1. 去除类型不匹配的 warning； 2. 将 NSString (QMUI) qmui_timeStringFromSeconds 重命名为 qmui_timeStringWithMinsAndSecsFromSecs； 3. 修复 UIControl (QMUI) 被废弃的 qmui_needsTakeOverTouchEvent 无法使用的问题

+++ QMUIKit/UIComponents/AssetLibrary/QMUIAsset.m
@@ -284,9 +284,9 @@
         if (completion) {
             [self assetSize:^(long long size) {
                 // 获取 NSData 数据
-                uint8_t *buffer = malloc(size);
+                uint8_t *buffer = malloc((size_t)size);
                 NSError *error;
-                NSUInteger bytes = [_alAssetRepresentation getBytes:buffer fromOffset:0 length:size error:&error];
+                NSUInteger bytes = [_alAssetRepresentation getBytes:buffer fromOffset:0 length:(NSUInteger)size error:&error];
                 NSData *imageData = [NSData dataWithBytes:buffer length:bytes];
                 free(buffer);
                 // 判断是否为 GIF 图
